revision des model à noeud et du type szt

+++ entities/typemps.h
@@ -9,7 +9,7 @@
 namespace typeMPS {
 using idt = unsigned;
 //using codeType = unsigned;
-using szt = unsigned;
+using szt = long unsigned;
 
 class flag {
 public:
